requirements.txt
PyYAML~=6.0.1
requests~=2.31.0
responses~=0.25.0
setuptools~=68.2.0